IOT-154: Fix package names to comply with the pattern com.hortonworks.iotas
