+++ core/src/test/java/com/hortonworks/iotas/util/FileSystemJarStorageTest.java
@@ -1,0 +1,11 @@
+package com.hortonworks.iotas.util;
+
+import com.hortonworks.iotas.util.FileSystemJarStorage;
+import com.hortonworks.iotas.util.JarStorage;
+
+public class FileSystemJarStorageTest extends AbstractJarStorageTest {
+    @Override
+    public JarStorage getJarStorage() {
+        return new FileSystemJarStorage();
+    }
+}
