--- storm/src/main/java/com/hortonworks/bolt/PrinterBolt.java
@@ -1,18 +1,0 @@
-package com.hortonworks.bolt;
-
-import org.apache.storm.topology.BasicOutputCollector;
-import org.apache.storm.topology.OutputFieldsDeclarer;
-import org.apache.storm.topology.base.BaseBasicBolt;
-import org.apache.storm.tuple.Tuple;
-
-public class PrinterBolt extends BaseBasicBolt {
-  @Override
-  public void execute(Tuple tuple, BasicOutputCollector collector) {
-    System.out.println(tuple);
-  }
-
-  @Override
-  public void declareOutputFields(OutputFieldsDeclarer ofd) {
-  }
-
-}
